0day: fix -Wformat warning in namespace_devs.c
0day says:
>> include/linux/dynamic_debug.h:64:16: warning: format '%lld' expects argument of type 'long long int', but argument 5 has type 'resource_size_t' [-Wformat=]

   428          if (val > allocated && val > min_available) {
   429                  if (min_dimm)
 > 430                          dev_dbg(dev, "%s only has %lld bytes available\n",
   431                                          dev_name(&min_dimm->dev),
   432                                          min_available);

Reported-by: kbuild test robot <fengguang.wu@intel.com>
Tested-by: Ross Zwisler <ross.zwisler@linux.intel.com>

+++ drivers/block/nd/namespace_devs.c
@@ -505,7 +505,8 @@ static ssize_t __size_store(struct device *dev, const char *buf)
 
 	if (val % nd_region->ndr_mappings)
 		dev_info(dev, "%lld truncated to %lld setting namespace size\n",
-				val, (val / nd_region->ndr_mappings)
+				val, (unsigned long long)
+				(val / nd_region->ndr_mappings)
 				* nd_region->ndr_mappings);
 	val /= nd_region->ndr_mappings;
 	allocated /= nd_region->ndr_mappings;
